split out options test
